# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Carbon Credit & Stock Tracker API
  description: Comprehensive API for tracking carbon credits, stocks, and generating ML-powered forecasts
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UNFCCCEmissions:
      type: object
      properties:
        party_code:
          type: string
          description: Country code
        gases:
          type: string
          description: Greenhouse gases queried
        data:
          type: array
          items:
            type: object
            properties:
              year:
                type: integer
              gas:
                type: string
              value:
                type: number
              category_name:
                type: string
        summary:
          type: object
          properties:
            total_records:
              type: integer
            years:
              type: array
              items:
                type: integer
            categories:
              type: array
              items:
                type: string
...
paths:
...
  /unfccc/status:
    get:
      summary: Get UNFCCC service status
      tags: [UNFCCC]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Service status information
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                  python_package_installed:
                    type: boolean
                  api_initialized:
                    type: boolean
                  zenodo_available:
                    type: boolean
  /unfccc/parties:
    get:
      summary: Get available parties (countries)
      tags: [UNFCCC]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of available parties
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    name:
                      type: string
                    type:
                      type: string
  /unfccc/emissions/{partyCode}:
    get:
      summary: Get emissions data for a specific party
      tags: [UNFCCC]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: partyCode
          required: true
          schema:
            type: string
          description: ISO code of the party (e.g., USA, GBR, DEU)
        - in: query
          name: gases
          schema:
            type: string
          description: Comma-separated list of gases (e.g., CO2,CH4,N2O)
      responses:
        '200':
          description: Emissions data for the party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UNFCCCEmissions'
        '400':
          description: Invalid party code
        '500':
          description: Internal server error
  /unfccc/summary:
    post:
      summary: Get emissions summary for multiple parties
      tags: [UNFCCC]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - partyCodes
              properties:
                partyCodes:
                  type: array
                  items:
                    type: string
                  description: List of party codes
      responses:
        '200':
          description: Emissions summary for multiple parties
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                  total_parties:
                    type: integer
                  successful_queries:
                    type: integer
...